[ADF-5174] Add Permisssions Dialog to adf-testing Lib
