+++ lib/testing/src/lib/content-services/dialog/public-api.ts
@@ -16,4 +16,5 @@
  */
 
 export * from './content-node-selector-dialog.page';
+export * from './add-permissions-dialog.page';
 export * from './download-dialog.page';
